[update] 수정시 사용할 useEffect 내부로 변경, 이미지 삭제버튼 통합 #103
