+++ src/features/exhibition/ExhibitionForm.jsx
@@ -2,47 +2,39 @@ import styled from "styled-components";
 import React, { useEffect } from "react";
 import { MdOutlineFileDownload } from "react-icons/md";
 import { Flex } from "../../components/Flex";
-import {
-  useGetPostimgurlEx,
-  useGetimgurlEx,
-} from "../../hooks/exhibition/useGetimgurlEx";
-import {
-  useDropzoneinputEx,
-  useDropzoneinputPostEx,
-} from "../../hooks/exhibition/useDropzoneEx";
 import { useSetExhibition } from "../../hooks/exhibition/useSetExhibition";
 import { useGetImgUrl } from "./CreateURL";
 import { useDropzoneInput } from "../../hooks/exhibition/useDropZone";
 
 function ExhibitionForm(props) {
   const info = props.Detaildata?.exhibitionInfo;
-  const ExAddress = props.Detaildata?.exhibitionInfo.ExhibitionAddress;
   const sourceUrl = "exhibition";
   const [
-    exhibition,
-    setExhibition,
-    exhibitionKind,
-    changeOnOff,
-    authorid,
-    authorName,
-    setAuthorName,
-    handleClick,
-    onchangeHandler,
-    setExhibitionKind,
-  ] = useSetExhibition();
-
-  const [postfiles, setPostFiles, getRootPropsPOST, getInputPropsPOST] =
-    useDropzoneInput(1);
-  const [files, setFiles, getRootProps, getInputProps] = useDropzoneInput(10);
+    postfiles,
+    setPostFiles,
+    getRootPropsPOST,
+    getInputPropsPOST,
+    deleteImgPOST,
+  ] = useDropzoneInput(1);
+  const [files, setFiles, getRootProps, getInputProps, deleteImg] =
+    useDropzoneInput(10);
   const [s3ImgUrlHandle] = useGetImgUrl(sourceUrl);
   const [s3PostImgUrlHandle] = useGetImgUrl(sourceUrl, true);
-  //파일이 언마운트 되면 리샛시켜줌
-  useEffect(() => {
-    return () => {
-      files.forEach((file) => URL.revokeObjectURL(file.preview));
-      postfiles.forEach((file) => URL.revokeObjectURL(file.preview));
-    };
-  }, []);
+  const [
+    exhibition,
+    exhibitionKind,
+    changeOnOff,
+    authorName,
+    handleClick,
+    onchangeHandler,
+  ] = useSetExhibition(
+    props.DetailLoading,
+    props.DetailError,
+    props.Detaildata,
+    setFiles,
+    setPostFiles
+  );
+
   //제출버튼
   const submitHandler = (event) => {
     event.preventDefault();
@@ -55,22 +47,8 @@ function ExhibitionForm(props) {
       exhibitionKind,
     });
   };
-  //이미지 삭제버튼
-  const deletePostImg = (name, index) => {
-    if (name === "postFile") {
-      const currentFiles = [...postfiles];
-      URL.revokeObjectURL(currentFiles.preview);
-      currentFiles.splice(index, 1);
-      setPostFiles(currentFiles);
-    }
-    if (name === "files") {
-      const currentFiles = [...files];
-      URL.revokeObjectURL(currentFiles.preview);
-      currentFiles.splice(index, 1);
-      setFiles(currentFiles);
-    }
-  };
   //업데이트 버튼
+
   const submitUpdateHandler = (event) => {
     event.preventDefault();
     let urls = null;
@@ -78,12 +56,12 @@ function ExhibitionForm(props) {
     if (!postfiles[0].type) {
       posturl = info.postImage;
     } else {
-      posturl = s3PostImgUrlHandle(sourceUrl);
+      posturl = s3PostImgUrlHandle(postfiles);
     }
     if (!files[0].type) {
       urls = info.ExhibitionImgs;
     } else {
-      urls = s3ImgUrlHandle(sourceUrl);
+      urls = s3ImgUrlHandle(files);
     }
     props.updateExhibition({
       ...exhibition,
@@ -92,67 +70,9 @@ function ExhibitionForm(props) {
       exhibitionKind,
     });
   };
-  //data가 들어오면 data를 넣어주는 effect
-  useEffect(() => {
-    // 서버에서 받아온 데이터가 로딩되면 exhibition state를 업데이트
-    //!나중에 필수값 유효성검사 필요!
-    //!작가,이미지 order없이 받는데 어떻게 해야 하는가?
-    if (!props.DetailLoading && !props.DetailError && props.Detaildata) {
-      //!value값 따로??
-      setAuthorName(info.ExhibitionAuthors[authorid].author);
-      const newarr = [...exhibition.authors];
-      newarr.splice(authorid, 1, {
-        order: authorid + 1,
-        author: info.ExhibitionAuthors[authorid].author,
-      });
-      const newExCodeArr = info.ExhibitionCategories.map(
-        (item) => item.categoryCode
-      );
-      //TODO 디테일한 유효성 검사 필요
-      setExhibition((prevExhibition) => ({
-        ...prevExhibition,
-        startDate: info.startDate.slice(0, 10), //
-        endDate: info.endDate.slice(0, 10), //
-        exhibitionTitle: info.exhibitionTitle, //
-        exhibitionEngTitle: info.exhibitionEngTitle, //
-        exhibitionDesc: info.exhibitionDesc, //
-        exhibitionHost: info.exhibitionHost, //
-        exhibitionCode: info.exhibitionStatus,
-        entranceFee: info.entranceFee, //
-        artWorkCnt: info.artWorkCnt, //
-        agencyAndSponsor: info.agencyAndSponsor, //
-        location: info.location, //
-        contact: info.contact, //
-        authors: newarr, //
-        exhibitionCategoty: newExCodeArr, //
-        detailLocation: {
-          //
-          zonecode: ExAddress.zonecode,
-          address: ExAddress.address,
-          addressEnglish: ExAddress.addressEnglish,
-          addressType: ExAddress.addressType,
-          buildingName: ExAddress.buildingName,
-          buildingCode: ExAddress.buildingCode,
-          roadAddress: ExAddress.roadAddress,
-          roadAddressEnglish: ExAddress.roadAddressEnglish,
-          autoJibunAddress: ExAddress.autoJibunAddress,
-          autoJibunAddressEnglish: ExAddress.autoJibunAddressEnglish,
-          roadname: ExAddress.roadname,
-          roadnameCode: ExAddress.roadnameCode,
-          roadnameEnglish: ExAddress.roadnameEnglish,
-        },
-      }));
-      setExhibitionKind(info.exhibitionKind);
-      //*썸네일 미리보기 가지고 와보기
-      setPostFiles([{ preview: info?.postImage }]);
-      //*일반 파일 미리보기 가지고 와보기
-      const previewFileArr = info?.ExhibitionImgs.map((file) => {
-        return { preview: file.imgUrl };
-      });
-      setFiles(previewFileArr);
-    }
-  }, [props.DetailLoading, props.DetailError, props.Detaildata]);
-  console.log("보내질 값", files);
+
+  console.log("외부파일 그냥파일", files);
+  console.log("외부파일포스트", postfiles);
   return (
     <Flex
       as={"form"}
@@ -194,7 +114,7 @@ function ExhibitionForm(props) {
                 <Postimg key={file.name} src={file.preview} />
                 <button
                   type="button"
-                  onClick={() => deletePostImg("postFile", index)}
+                  onClick={() => deleteImgPOST("postFile", index)}
                 >
                   삭제
                 </button>
@@ -436,7 +356,7 @@ function ExhibitionForm(props) {
                   </Thumb>
                   <button
                     type="button"
-                    onClick={() => deletePostImg("files", index)}
+                    onClick={() => deleteImg("files", index)}
                   >
                     삭제
                   </button>
