+++ src/hooks/exhibition/useSetExhibition.js
@@ -1,10 +1,16 @@
-import { useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { useDaumPostcodePopup } from "react-daum-postcode";
 
-export const useSetExhibition = () => {
+export const useSetExhibition = (
+  DetailLoading,
+  DetailError,
+  Detaildata,
+  setFiles,
+  setPostFiles
+) => {
+  const info = Detaildata?.exhibitionInfo;
+  const ExAddress = Detaildata?.exhibitionInfo.ExhibitionAddress;
   let authorid = 0;
-  const [exhibitionKind, setExhibitionKind] = useState("EK0001");
-  const [authorName, setAuthorName] = useState("");
   const templete = {
     startDate: "",
     endDate: "",
@@ -38,6 +44,8 @@ export const useSetExhibition = () => {
       roadnameEnglish: "",
     },
   };
+  const [exhibitionKind, setExhibitionKind] = useState("EK0001");
+  const [authorName, setAuthorName] = useState("");
   const [exhibition, setExhibition] = useState({ ...templete });
   //카카오 주소
   const open = useDaumPostcodePopup(process.env.REACT_APP_KAKAO_ADDRESS_URL);
@@ -169,17 +177,75 @@ export const useSetExhibition = () => {
       setExhibitionKind(name);
     }
   };
-
+  useEffect(() => {
+    // 서버에서 받아온 데이터가 로딩되면 exhibition state를 업데이트
+    //!나중에 필수값 유효성검사 필요!
+    //!작가,이미지 order없이 받는데 어떻게 해야 하는가?
+    if (!DetailLoading && !DetailError && Detaildata) {
+      //!value값 따로??
+      setAuthorName(info.ExhibitionAuthors[authorid].author);
+      const newarr = [...exhibition.authors];
+      newarr.splice(authorid, 1, {
+        order: authorid + 1,
+        author: info.ExhibitionAuthors[authorid].author,
+      });
+      const newExCodeArr = info.ExhibitionCategories.map(
+        (item) => item.categoryCode
+      );
+      //TODO 디테일한 유효성 검사 필요
+      setExhibition((prevExhibition) => ({
+        ...prevExhibition,
+        startDate: info.startDate.slice(0, 10), //
+        endDate: info.endDate.slice(0, 10), //
+        exhibitionTitle: info.exhibitionTitle, //
+        exhibitionEngTitle: info.exhibitionEngTitle, //
+        exhibitionDesc: info.exhibitionDesc, //
+        exhibitionHost: info.exhibitionHost, //
+        // exhibitionCode: info.exhibitionStatus,
+        entranceFee: info.entranceFee, //
+        artWorkCnt: info.artWorkCnt, //
+        agencyAndSponsor: info.agencyAndSponsor, //
+        location: info.location, //
+        contact: info.contact, //
+        authors: newarr, //
+        exhibitionCategoty: newExCodeArr, //
+        detailLocation: {
+          //
+          zonecode: ExAddress.zonecode,
+          address: ExAddress.address,
+          addressEnglish: ExAddress.addressEnglish,
+          addressType: ExAddress.addressType,
+          buildingName: ExAddress.buildingName,
+          buildingCode: ExAddress.buildingCode,
+          roadAddress: ExAddress.roadAddress,
+          roadAddressEnglish: ExAddress.roadAddressEnglish,
+          autoJibunAddress: ExAddress.autoJibunAddress,
+          autoJibunAddressEnglish: ExAddress.autoJibunAddressEnglish,
+          roadname: ExAddress.roadname,
+          roadnameCode: ExAddress.roadnameCode,
+          roadnameEnglish: ExAddress.roadnameEnglish,
+        },
+      }));
+      setExhibitionKind(info.exhibitionKind);
+      //*썸네일 미리보기 가지고 와보기
+      setPostFiles([{ preview: info?.postImage }]);
+      //*일반 파일 미리보기 가지고 와보기
+      const previewFileArr = info?.ExhibitionImgs.map((file) => {
+        return { preview: file.imgUrl };
+      });
+      setFiles(previewFileArr);
+    }
+    // clean-up 함수
+    return () => {
+      setExhibition({ ...templete });
+    };
+  }, [DetailLoading, DetailError, Detaildata]);
   return [
     exhibition,
-    setExhibition,
     exhibitionKind,
     changeOnOff,
-    authorid,
     authorName,
-    setAuthorName,
     handleClick,
     onchangeHandler,
-    setExhibitionKind,
   ];
 };
